+++ App/rosenFrontend/src/components/App.js
@@ -1,11 +1,9 @@
-import logo from "./logo.svg";
-import "./styles/App.css";
+import "../styles/App.css";
 
 function App() {
   return (
     <div className="App">
       <header className="App-header">
-        <img src={logo} className="App-logo" alt="logo" />
         <p>
           Edit <code>src/App.js</code> and save to reload.
         </p>
